Changed some value types
DERIVE is change over time, COUNTER is change over time, but does not allow negative values, but handles wrapping correctly, GAUGE is current value

RAM is a GAUGE, we are interested in the current value, not the change

NetworkIO is normaly a COUNTER, but when the router reboots the value resets to 0 and does not wrap. Reset to 0 is handled by setting min = 0. I assume wraps are rare and routerreboots (= resets) are common, so DERIVE with min = 0 is better than COUNTER.